requirements.txt
Flask==2.3.3
langdetect==1.0.9
PyPDF2==3.0.1
fpdf==1.7.2
arabic-reshaper==2.1.4
python-bidi==0.4.2
nltk==3.8.1
pandas==1.5.3
numpy==1.23.5
textdistance==4.5.0
transformers==4.35.0
torch==2.0.1
language_tool_python
rapidfuzz
pyspellchecker 
tensorflow
tf-keras
sentencepiece
sacremoses